Consolidate to just step.make_output_path
Since output is controlled by parameters, the basic logic of
pipeline.make_output_path is the same as the more basic
stop.make_output_path. Now the default is simply the
pipeline.make_output_path.

Besides reviewer modifications, this accomplished the desired task.

Resolves #678
